backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy>=2.0.36
psycopg2-binary>=2.9.9
bcrypt>=4.2.0
pyjwt>=2.9.0
python-multipart>=0.0.12
python-dotenv>=1.0.1
pydantic[email]>=2.10.0
Pillow>=10.4.0
